Add Avatar model and update UploadExcel form
The Avatar model has been added to `models.py` to manage user avatar images. This includes a new migration file (`0002_avatar.py`). In `forms.py`, a label and requirement have been added to the file upload in the UploadExcel form. Also, a new field 'exclude' is added to exclude 'user'.